requirements.txt
fastapi
uvicorn
python-dotenv
openai
httpx
aiofiles
python-multipart
google-generativeai
